fix: post-merge cleanup
