+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/exec/systemcontracts/hts/update/address_0x167/UpdateDecoderTest.java
@@ -14,14 +14,11 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.hts.update;
+package com.hedera.node.app.service.contract.impl.test.exec.systemcontracts.hts.update.address_0x167;
 
-import static com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x16c.UpdateTranslator.TOKEN_UPDATE_INFO_FUNCTION_WITH_METADATA;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.FUNGIBLE_TOKEN_HEADLONG_ADDRESS;
-import static com.hedera.node.app.service.contract.impl.test.TestHelpers.NON_FUNGIBLE_TOKEN_HEADLONG_ADDRESS;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.OWNER_HEADLONG_ADDRESS;
 import static com.hedera.node.app.service.contract.impl.test.TestHelpers.OWNER_ID;
-import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.BDDMockito.given;
@@ -29,9 +26,8 @@ import static org.mockito.BDDMockito.given;
 import com.esaulpaugh.headlong.abi.Tuple;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.AddressIdConverter;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.HtsCallAttempt;
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateDecoder;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateExpiryTranslator;
-import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.UpdateNFTsMetadataTranslator;
+import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x167.UpdateDecoder;
 import com.hedera.node.app.service.contract.impl.exec.systemcontracts.hts.update.address_0x167.UpdateTranslator;
 import java.time.Instant;
 import org.apache.tuweni.bytes.Bytes;
@@ -52,7 +48,6 @@ class UpdateDecoderTest {
 
     private final UpdateDecoder subject = new UpdateDecoder();
     private final String newName = "NEW NAME";
-    private final String metadata = "LionTigerBear";
     private static final long EXPIRY_TIMESTAMP = Instant.now().plusSeconds(3600).toEpochMilli() / 1000;
     private static final long AUTO_RENEW_PERIOD = 8_000_000L;
     private final Tuple expiry = Tuple.of(EXPIRY_TIMESTAMP, OWNER_HEADLONG_ADDRESS, AUTO_RENEW_PERIOD);
@@ -68,21 +63,6 @@ class UpdateDecoderTest {
             new Tuple[] {},
             // Expiry
             expiry);
-
-    private final Tuple hederaTokenWithMetadata = Tuple.from(
-            newName,
-            "symbol",
-            OWNER_HEADLONG_ADDRESS,
-            "memo",
-            true,
-            1000L,
-            false,
-            // TokenKey
-            new Tuple[] {},
-            // Expiry
-            expiry,
-            // Metadata,
-            metadata.getBytes());
 
     @Nested
     class UpdateTokenInfoAndExpiry {
@@ -126,17 +106,6 @@ class UpdateDecoderTest {
         }
 
         @Test
-        void updateWithMetadataWorks() {
-            final var encoded = Bytes.wrapByteBuffer(TOKEN_UPDATE_INFO_FUNCTION_WITH_METADATA.encodeCallWithArgs(
-                    FUNGIBLE_TOKEN_HEADLONG_ADDRESS, hederaTokenWithMetadata));
-            given(attempt.input()).willReturn(encoded);
-
-            final var body = subject.decodeTokenUpdateWithMetadata(attempt);
-            final var tokenUpdate = body.tokenUpdateOrThrow();
-            assertEquals(tokenUpdate.metadata().asUtf8String(), metadata);
-        }
-
-        @Test
         void updateExpiryV1Works() {
             final var encoded =
                     Bytes.wrapByteBuffer(UpdateExpiryTranslator.UPDATE_TOKEN_EXPIRY_INFO_V1.encodeCallWithArgs(
@@ -171,19 +140,5 @@ class UpdateDecoderTest {
             assertEquals(AUTO_RENEW_PERIOD, tokenUpdate.autoRenewPeriod().seconds());
             assertEquals(OWNER_ID, tokenUpdate.autoRenewAccount());
         }
-    }
-
-    @Test
-    void updateNFTsMetadataWorks() {
-        final var encoded = Bytes.wrapByteBuffer(UpdateNFTsMetadataTranslator.UPDATE_NFTs_METADATA.encodeCallWithArgs(
-                NON_FUNGIBLE_TOKEN_HEADLONG_ADDRESS, new long[] {1, 2, 3}, "Jerry".getBytes()));
-        given(attempt.input()).willReturn(encoded);
-
-        final var body = subject.decodeUpdateNFTsMetadata(attempt);
-        final var tokenUpdate = requireNonNull(body).tokenUpdateNftsOrThrow();
-
-        assertNotNull(tokenUpdate.metadata());
-        assertEquals("Jerry", tokenUpdate.metadata().asUtf8String());
-        assertEquals(3, tokenUpdate.serialNumbers().size());
     }
 }
